# docs/swagger/swagger.yaml
definitions:
...
  model.HTTPError:
    properties:
      error_code:
        type: string
      error_msg:
        type: string
    type: object
...
  model.ServerStatus:
    properties:
      cluster_id:
        type: string
      git_hash:
        type: string
      id:
        type: string
      is_owner:
        type: boolean
      liveness:
        type: integer
      pid:
        type: integer
      version:
        type: string
    type: object
...
  v2.ServerStatus:
    properties:
      cluster_id:
        type: string
      git_hash:
        type: string
      id:
        type: string
      is_owner:
        type: boolean
      liveness:
        type: integer
      pid:
        type: integer
      version:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/status:
    get:
      consumes:
      - application/json
      description: get the status of a server(capture)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ServerStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
      summary: Get server status
      tags:
      - common
...
  /api/v2/status:
    get:
      consumes:
      - application/json
      description: This API is a synchronous interface. If the request is successful,
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.ServerStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
      summary: Get the status information of a TiCDC node
      tags:
      - common
      - v2
swagger: "2.0"